feat(api): generated reflection tables, and the field that was missing
gen_cpp.luau turns the IDL into the ClassRegistry's descriptors: 13 classes,
17 properties, 43 methods, 14 events, 35 native functions, 0 skipped, 0
unbacked. The declared set and the hand-written native_accessors.cpp agree name
for name, 35 for 35, which is the check that matters -- a mismatch there is a
link error rather than a wrong answer at runtime.

The generator found a hole I had left. World::create attaches components through
ClassDescriptor hooks, and the IDL had no way to say which class owns which
component, so the generator would have emitted nullptr for every one: a
generated Part would have been born with no PartComponent and every BasePart
accessor would have seen a null pool. It reported that instead of inventing an
IDL field, which is exactly right. Schema.Class gains `Storage`, a name rather
than a storage expression, mirroring Property.Native -- BasePart, Model and
Script carry it, and Part deliberately does not, because create walks the
ancestry root-first and declaring it twice would attach twice.

Doc keys are emitted empty on purpose. Nothing formats one yet: the docs
pipeline is M3, and generating 76 catalog entries for text nothing reads, purely
to quiet a lint, is inventing work. The generated banner says so, so the empty
key reads as a decision.

Freshness-gated like the type definitions, and proved by tampering: appending
one line to a generated file turns the gate red, and it names both causes
because both have the same fix.

Also fixes main, which my previous commit had left red: three doctest::Approx
calls on f32 promoted to double, which MSVC accepts and Clang's
-Wdouble-promotion does not. Found by the generator agent running the Linux
tier, not by me running it before committing.

+++ engine/core/tests/math_tests.cpp
@@ -729,7 +729,8 @@ TEST_CASE("euler YXZ round-trips through the rotation it builds")
                 for (int column = 0; column < 3; ++column)
                 {
                     for (int row = 0; row < 3; ++row)
-                        CHECK(built.m[column][row] == doctest::Approx(rebuilt.m[column][row]).epsilon(1e-4));
+                        CHECK(static_cast<f64>(built.m[column][row])
+                              == doctest::Approx(static_cast<f64>(rebuilt.m[column][row])).epsilon(1e-4));
                 }
             }
         }
@@ -741,20 +742,21 @@ TEST_CASE("euler YXZ keeps pitch in the principal range and resolves the poles")
     // Pitch is the middle rotation, so its branch is the one that has to be
     // chosen: [-pi/2, pi/2] is the documented half.
     const Vec3 recovered = toEulerYxz(fromEulerYxz(Vec3{2.0f, 0.3f, 0.4f}));
-    CHECK(recovered.x <= doctest::Approx(1.5708).epsilon(1e-4));
-    CHECK(recovered.x >= doctest::Approx(-1.5708).epsilon(1e-4));
+    CHECK(static_cast<f64>(recovered.x) <= doctest::Approx(1.5708).epsilon(1e-4));
+    CHECK(static_cast<f64>(recovered.x) >= doctest::Approx(-1.5708).epsilon(1e-4));
 
     // At the pole, yaw and roll describe the same rotation and the pair is not
     // recoverable. Roll resolves to zero rather than splitting the angle
     // arbitrarily, which is what keeps a round trip stable.
     const Vec3 atPole = toEulerYxz(fromEulerYxz(Vec3{1.5707963f, 0.8f, 0.6f}));
-    CHECK(atPole.z == doctest::Approx(0.0f).epsilon(1e-4));
+    CHECK(static_cast<f64>(atPole.z) == doctest::Approx(0.0).epsilon(1e-4));
 
     const Mat3 built = fromEulerYxz(Vec3{1.5707963f, 0.8f, 0.6f});
     const Mat3 rebuilt = fromEulerYxz(atPole);
     for (int column = 0; column < 3; ++column)
     {
         for (int row = 0; row < 3; ++row)
-            CHECK(built.m[column][row] == doctest::Approx(rebuilt.m[column][row]).epsilon(1e-3));
+            CHECK(static_cast<f64>(built.m[column][row])
+                  == doctest::Approx(static_cast<f64>(rebuilt.m[column][row])).epsilon(1e-3));
     }
 }
